server/requirements.txt
python-telegram-bot>=21.0
scrapy>=2.11
openai>=1.0
python-dotenv>=1.0
aiohttp>=3.9
beautifulsoup4>=4.12
lxml>=5.0
google-search-results>=2.4
gspread>=6.0
google-auth>=2.0
